Razor class library the assets.
