+++ gulpfile.js
@@ -1,5 +1,5 @@
 /// <binding ProjectOpened='default' />
-const { watch, src, dest } = require('gulp');
+const { watch, src, dest, on } = require('gulp');
 const del = require('del');
 var concat = require('gulp-concat');
 var uglify = require('gulp-uglify');
@@ -7,65 +7,38 @@ var sourcemaps = require('gulp-sourcemaps');
 const cleanCSS = require('gulp-clean-css');
 const transformManifest = require('./gulp-transformManifest');
 
-const sourceFolders = [
-    'uSync.BackOffice.Assets/App_Plugins/'
-];
+const version = "9.2.0";
 
-const pluginName = 'uSync';
+const scriptSource = 'uSync.Assets/scripts';
+const styleSource = 'uSync.Assets/css';
+
+// we still need to do this on dev to copy the lang files.
+const pluginSource = 'uSync.Assets/App_Plugins/';
+
 const destination = 'uSync.Site/App_Plugins/';
-const minDest = 'build-files/App_Plugins/';
+const minDest = 'uSync.Assets/wwwroot';
 
-// fetch command line arguments
-const arg = (argList => {
-
-    let arg = {}, a, opt, thisOpt, curOpt;
-    for (a = 0; a < argList.length; a++) {
-  
-      thisOpt = argList[a].trim();
-      opt = thisOpt.replace(/^\-+/, '');
-  
-      if (opt === thisOpt) {
-  
-        // argument value
-        if (curOpt) arg[curOpt] = opt;
-        curOpt = null;
-  
-      }
-      else {
-  
-        // argument name
-        curOpt = opt;
-        arg[curOpt] = true;
-  
-      }
-  
-    }
-  
-    return arg;
-  
-  })(process.argv);
-
-  function copy(path, baseFolder) {
+function copy(path, baseFolder) {
     return src(path, { base: baseFolder })
         .pipe(dest(destination));
 }
 
-function minifyJs(path, version) {
+function minifyJs(path) {
 
     return src(path, { base: path })
         .pipe(sourcemaps.init())
             .pipe(concat('usync.' + version + '.min.js'))
             .pipe(uglify({ mangle: false }))
         .pipe(sourcemaps.write('.'))
-        .pipe(dest(minDest + pluginName));
+        .pipe(dest(minDest));
 }
 
 
-function minifyCss(path, version) {
+function minifyCss(path) {
     return src(path, { base: path })
         .pipe(cleanCSS({ compatibility: 'ie8'}))
         .pipe(concat('usync.' + version + '.min.css'))
-        .pipe(dest(minDest + pluginName));
+        .pipe(dest(minDest));
 }
 
 function copyRequired(path, baseFolder, destFolder) {
@@ -77,7 +50,8 @@ function time() {
     return '[' + new Date().toISOString().slice(11, -5) + ']';
 }
 
-function updateManifest(manifest, version) {
+/*
+function updateManifest(manifest) {
 
     return src(manifest)
         .pipe(transformManifest({
@@ -88,42 +62,39 @@ function updateManifest(manifest, version) {
         .pipe(dest(minDest + pluginName));
 
 }
+*/
 
+function doMinify(cb)
+{
+    let jsfiles = scriptSource + '**/*.js';
+    let cssFiles = styleSource + '**/*.css';
 
-////////////
-exports.default = function() {
-    sourceFolders.forEach(function (sourceFolder) {
-
-        let source = sourceFolder + '**/*';
-       
-        watch(source, { ignoreInitial: false })
-            .on('change', function (path, stats) {
-                console.log(time(), path, sourceFolder, 'changed');
-                copy(path, sourceFolder);
-            })
-            .on('add', function (path, stats) {
-                console.log(time(), path, sourceFolder, 'added');
-                copy(path, sourceFolder);
-            });
-    });
-
-};
-
-exports.minify = function (cb) {
-    
-    var version = arg.release;
-   
-    sourceFolders.forEach(function (sourceFolder) {
-        let source = sourceFolder + '**/*';
-        let jsfiles = sourceFolder + '**/*.js';
-        let cssFiles = sourceFolder + '**/*.css';
-
-        minifyJs(jsfiles, version);
-        minifyCss(cssFiles, version);
-        updateManifest(sourceFolder + pluginName + "/package.manifest", version);
-        copyRequired(source, sourceFolder, minDest);
-    });
-
-    cb();
+    minifyJs(jsfiles);
+    minifyCss(cssFiles);
 }
 
+////////////
+exports.default = function() 
+{
+        let source = pluginSource + '**/*';
+
+        // let jsfiles = scriptSource + '**/*.js';
+        // let cssFiles = styleSource + '**/*.css';
+        
+        watch(source, { ignoreInitial: false })
+            .on('change', function (path, stats) {
+                console.log(time(), path, 'changed');
+                copy(path, pluginSource);
+                doMinify();
+            })
+            .on('add', function (path, stats) {
+                console.log(time(), path, 'processed');
+                copy(path, pluginSource);
+                doMinify();
+            });
+}
+
+exports.minify = function (cb) {
+    doMinify();
+    cb();
+};
