+++ uSync.Assets/App_Plugins/uSync/backoffice/usync/dashboard.html
@@ -1,0 +1,17 @@
+﻿<div ng-controller="uSyncSettingsDashboardController as vm">
+    <umb-editor-view footer="false">
+        <umb-editor-header name="vm.page.title"
+                           description="vm.page.description"
+                           hide-description="false"
+                           name-locked="true"
+                           description-locked="true"
+                           hide-alias="true"
+                           hide-icon="true"
+                           navigation="vm.page.navigation"
+                           on-select-navigation-item="vm.selectNavigationItem(item)">
+        </umb-editor-header>
+        <umb-editor-container class="form-horizontal">
+            <umb-editor-sub-views sub-views="vm.page.navigation" model="vm"></umb-editor-sub-views>
+        </umb-editor-container>
+    </umb-editor-view>
+</div>
