+++ uSync.Assets/wwwroot/config/appsettings-usync-schema.json
@@ -1,0 +1,88 @@
+﻿{
+  "title": "uSyncSchema",
+  "$schema": "http://json-schema.org/draft-04/schema",
+  "type": "object",
+  "description": "the uSync settings",
+  "definitions": {
+    "uSync": {
+      "description": "Configuration of uSync syncing tool for Umbraco",
+      "properties": {
+        "Settings": {
+          "type": "object",
+          "description": "Default settings for uSync",
+          "properties": {
+            "RootFolder": {
+              "type": "string",
+              "description": "Location where all uSync files are saved by default",
+              "default": "uSync/v9"
+            },
+            "DefaultSet": {
+              "type": "string",
+              "description": "",
+              "default": "default"
+            },
+            "ImportAtStartup": {
+              "type": "string",
+              "description": "Import when Umbraco boots (can be group name or 'All' so everything is done, blank or 'none' == off)",
+              "default": "None"
+            },
+            "ExportAtStartup": {
+              "type": "string",
+              "description": "Export when Umbraco boots",
+              "default": "none"
+            },
+            "ExportOnSave": {
+              "type": "string",
+              "description": "Export when an item is saved in Umbraco",
+              "default": "All"
+            },
+            "UIEnabledGroups": {
+              "type": "string",
+              "description": "The handler groups that are enabled in the UI.",
+              "default": "All"
+            },
+            "ReportDebug": {
+              "type": "boolean",
+              "description": "Debug reports (creates an export into a temp folder for comparison)",
+              "default": false
+            },
+            "AddOnPing": {
+              "type": "boolean",
+              "description": "Ping the AddOnUrl to get the json used to show the addons dashboard",
+              "default": true
+            },
+            "FailOnMissingParent": {
+              "type": "boolean",
+              "description": "fail if the items parent is not in umbraco or part of the batch being imported",
+              "default": false
+            },
+            "CacheFolderKeys": {
+              "type": "boolean",
+              "description": "Cache keys lookups etc (performance)",
+              "default": true
+            },
+            "ShowVersionCheckWarning": {
+              "type": "boolean",
+              "description": " Show a version check warning to the user if the folder version is less than the version expected by uSync.",
+              "default": true
+            },
+            "CustomMappings": {
+              "type": "object",
+              "description": "Custom mapping keys, to add a simple config mapping to make one property behave like and existing one"
+            },
+            "SignalRRoot": {
+              "type": "string",
+              "description": "Location of the SignalR hub script",
+              "default": "/"
+            },
+            "EnableHistory": {
+              "type": "boolean",
+              "description": "",
+              "default": true
+            }
+          }
+        }
+      }
+    }
+  }
+}
