--- uSync.Site/App_Plugins/uSync/package.manifest
@@ -1,22 +1,0 @@
-{
-  "javascript": [
-
-    "~/App_Plugins/uSync/usync_903.js",
-
-    "~/App_Plugins/uSync/usync.service.js",
-    "~/App_Plugins/uSync/usync.hub.js",
-
-    "~/App_Plugins/uSync/backoffice/usync/usync.dashboard.controller.js",
-    "~/App_Plugins/uSync/settings/settings.controller.js",
-    "~/App_Plugins/uSync/settings/expansion.controller.js",
-    "~/App_Plugins/uSync/settings/usync.controller.js",
-
-    "~/App_Plugins/uSync/changedialog.controller.js",
-    "~/App_Plugins/uSync/components/usync.reportview.component.js",
-    "~/App_Plugins/uSync/components/usync.progressview.component.js"
-  ],
-  "css": [
-    "~/App_Plugins/uSync/usync_903.css",
-    "~/App_Plugins/uSync/usync.css"
-  ]
-}
