--- uSync.Site/App_Plugins/uSync/settings/default.html
@@ -1,102 +1,0 @@
-<div ng-controller="uSync8Controller as vm">
-
-
-    <div ng-if="vm.versionLoaded && vm.versionInfo.isCurrent != true" class="alert alert-info usync-alert">
-        <localize key="usync_newer"></localize> [v{{vm.versionInfo.versionInfo.Core}}]
-        <span ng-if="vm.versionInfo.versionInfo.message">
-            <a href="{{vm.versionInfo.versionInfo.link}}" target="_blank">
-                {{vm.versionInfo.versionInfo.message}}
-            </a>
-        </span>
-    </div>
-
-    <div class="usync-group-actions" ng-if="!vm.loading">
-        <umb-box class="usync-group-box" ng-repeat="group in vm.groups">
-            <umb-box-content>
-
-                <div class="usync-group-box-title">
-                    <i class="icon {{group.icon}}"></i>
-                    <h2>{{group.name}}</h2>
-                </div>
-                <small><localize key="usync_{{group.key}}-description"></small>
-
-                <div class="usync-group-buttons">
-                    <umb-button button-style="action" action="vm.report(group.name)" label="Report"></umb-button>
-
-                    <umb-button-group default-button="vm.importGroup[group.name].defaultButton"
-                                      sub-buttons="vm.importGroup[group.name].subButtons"
-                                      direction="down"
-                                      float="left"
-                                      button-style="success"
-                                      state="vm.importButton.state"></umb-button-group>
-                    <umb-button button-style="info" action="vm.exportGroup(group.name)" label="Export"></umb-button>
-                </div>
-            </umb-box-content>
-        </umb-box>
-
-        <umb-box class="usync-group-box" ng-if="vm.showEverything">
-            <umb-box-content>
-
-                <div class="usync-group-box-title">
-                    <i class="icon icon-paper-plane-alt color-deep-orange"></i>
-                    <h2><localize key="usync_everything"></localize></h2>
-                </div>
-                <small><localize key="usync_everything-description"></localize></small>
-
-                <div class="usync-group-buttons">
-                    <umb-button button-style="action" action="vm.report('')" label="Report"></umb-button>
-                    <umb-button-group default-button="vm.importButton.defaultButton"
-                        sub-buttons="vm.importButton.subButtons" direction="down" float="left" button-style="success"
-                        state="vm.importButton.state"></umb-button-group>
-                    <umb-button-group default-button="vm.exportButton.defaultButton"
-                        sub-buttons="vm.exportButton.subButtons" direction="down" float="left" button-style="info"
-                        state="vm.exportButton.state"></umb-button-group>
-                </div>
-            </umb-box-content>
-        </umb-box>
-    </div>
-
-
-    <div ng-if="vm.canHaveForms && !vm.reported && !vm.working" class="flex justify-center">
-        <localize key="usync_forms"></localize>
-    </div>
-
-    <umb-box ng-if="!vm.showSpinner && (vm.working || vm.reported)">
-        <umb-box-content>
-            <usync-progress-view update="vm.update" status="vm.status" hide-labels="false">
-            </usync-progress-view>
-            <div ng-if="vm.perf > 0" class="text-center muted">{{vm.perf | number: 0 }}ms</div>
-        </umb-box-content>
-        <umb-box-content>
-            <div ng-if="vm.warnings.message.length > 0" class="usync-warning usync-warning-{{vm.warnings.type}}">
-                <span ng-bind-html="vm.warnings.message"></span>
-            </div>
-        </umb-box-content>
-    </umb-box>
-
-    <div ng-if="vm.showSpinner && vm.working">
-        <umb-load-indicator></umb-load-indicator>
-    </div>
-
-
-
-    <div ng-if="vm.reported">
-        <usync-report-view actions="vm.actions" 
-                           results="vm.results" 
-                           hide-action="true" 
-                           hide-link="vm.hideLink"></usync-report-view>
-    </div>
-
-    <div class="usync-fresh-view" ng-if="vm.fresh">
-        <umb-empty-state>
-            
-            <div class="usync-group-box-title">
-                <i class="icon icon-infinity"></i>
-                <h2>uSync all the things</h2>
-            </div>
-            
-        </umb-empty-state>
-    </div>
-
-
-</div>
