--- uSync.Site/App_Plugins/uSync/settings/settings.overlay.html
@@ -1,11 +1,0 @@
-<div>
-    <p>
-        This is all the current settings in a format that can be used in the appsettings.json file
-    <p>
-    <p>
-        <strong>Note many of these values are default and do not need to be included in the appsettings.json file.</strong>
-        <a href="{{model.docslink}}" target="blank">see our  docs for more info</a>
-    </p>
-
-    <umb-code-snippet language="'json'">{{model.content}}</umb-code-snippet>
-</div>
