+++ uSync.Site/appsettings.json
@@ -1,7 +1,7 @@
 {
   "$schema": "./umbraco/config/appsettings-schema.json",
   "ConnectionStrings": {
-    "umbracoDbDSN": "Server=DESKTOP-2AFFJ8C\\SQLEXPRESS;Database=uSyncCore;Integrated Security=true"
+    "umbracoDbDSN": "Data Source=(localdb)\\MSSQLLocalDB;AttachDbFilename=|DataDirectory|\\Umbraco.mdf;Integrated Security=True"
   },
   "Serilog": {
     "MinimumLevel": {
@@ -27,7 +27,7 @@
         "UmbracoPath": "~/umbraco",
         "TimeOutInMinutes": 20,
         "UseHttps": false,
-        "Id": "6cac6005-516e-488c-a430-d8aca2efc5e0"
+        "Id": "b34520d0-6e89-4aef-b7cb-7ceebda6b03b"
       },
       "Hosting": {
         "Debug": true
